Add PALLOC patch for linux-4.4 kernel
I think that the original buffer of palloc ignores the migratetype of
pages and it may be important. Non-movable pages, to which most
allocations of the core kernel belong, have a fixed position in memory
and cannot be moved anywhere else. Movable pages can be moved around as
desired and pages that belong to userspace application pertain to this
category. So I calculated the page type in the original buffer, where
the page type is really different. In this way, when a running userspace
application wants to move the original pages to different pages, OS may
occur kernel panic if a part of these original pages are non-movable
pages.

Hence, I have extended the original buffer to multiple buffer lists,
each of which represents a migratetype, to prevent this problem.

Signed-off-by: Yupeng Li <ypli15@lzu.edu.cn>